Finished GameWorld
Added graphics, added a world, added collisions, added a bunch more monsters, made the last two guards not friendly (as a suprise!) added world collisions, added victory condition, reverted GameManager. It didn't function properly in the switch because it continued to load the scene over and over... Everything is working as intended. I am happy with the result. I learned a lot of tricks for next time!